Refactor interacción DB
